migrate deploy script to dedicated vps instead of personal server

+++ bin/deploy.sh
@@ -4,12 +4,16 @@ set -e
 set -x
 
 # Get connection details
-eval $(docker-machine env core2)
+eval $(docker-machine env aelf)
 
 # Deploy
 docker build -t aelf-api .
-docker stop aelf-api
-docker rm aelf-api
+
+if docker ps | grep -q aelf-api
+then
+    docker stop aelf-api
+    docker rm aelf-api
+fi
 docker run --name aelf-api -d -p 4001:4000 --restart always aelf-api
 
 # Follow
